rename file contents
