+++ pom.xml
@@ -1,11 +1,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.github.obsproth.obspassword</groupId>
-  <artifactId>obspassword</artifactId>
+  <groupId>com.github.obsproth.prismkey</groupId>
+  <artifactId>prismkey</artifactId>
   <packaging>jar</packaging>
   <version>1.0-SNAPSHOT</version>
-  <name>obspassword</name>
+  <name>prismkey</name>
   <url>http://maven.apache.org</url>
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -23,7 +23,7 @@
   </dependencies>
   <!-- プラグインの設定 -->
   <build>
-    <sourceDirectory>src</sourceDirectory>
+    <sourceDirectory>src/main</sourceDirectory>
       <plugins>
         <plugin>
             <groupId>org.apache.maven.plugins</groupId>
@@ -41,14 +41,14 @@
           <artifactId>maven-assembly-plugin</artifactId>
           <version>3.0.0</version>
           <configuration>
-            <finalName>obspassword</finalName>
+            <finalName>prismkey</finalName>
             <descriptorRefs>
               <!-- 依存するリソースをすべてjarに同梱する -->
               <descriptorRef>jar-with-dependencies</descriptorRef>
             </descriptorRefs>
             <archive>
               <manifest>
-                <mainClass>com.github.obsproth.obspassword.cli.ObsPassword</mainClass>
+                <mainClass>com.github.obsproth.prismkey.cli.PrismKeyCLI</mainClass>
               </manifest>
             </archive>
           </configuration>
